Integrity checks in tests.
