+++ app/test/shared/test_services.dart
@@ -31,6 +31,7 @@ import 'package:pub_dartlang_org/shared/email.dart';
 import 'package:pub_dartlang_org/shared/exceptions.dart'
     show AuthorizationException;
 import 'package:pub_dartlang_org/shared/handler_helpers.dart';
+import 'package:pub_dartlang_org/shared/integrity.dart';
 import 'package:pub_dartlang_org/shared/popularity_storage.dart';
 import 'package:pub_dartlang_org/shared/redis_cache.dart';
 import 'package:pub_dartlang_org/search/search_client.dart';
@@ -100,6 +101,13 @@ void testWithServices(String name, Future fn()) {
 
           await fork(() async {
             await fn();
+            // post-test integrity check
+            final problems =
+                await IntegrityChecker(dbService).check(ignorePackages: true);
+            if (problems.isNotEmpty) {
+              throw Exception(
+                  '${problems.length} integrity problems detected. First: ${problems.first}');
+            }
           });
         });
       });
